requirements.txt
psycopg2-binary==2.9.9
Pillow==10.2.0
openpyxl==3.1.2
python-docx==1.1.0
